item45: make `alphabetize` robust to non-BMP data
Summary:
See <https://wchargin.github.io/posts/encodings-and-assumptions> for
explanation, timing data, etc. The gist of it is that because Java
strings are encoded with UTF-16, the multiset of `char` values of a
string is not in bijection with the multiset of Unicode code points in
that string. The strings "U+10402 U+10C04" and "U+10404 U+10C02" clearly
are not anagrams, yet have the same sorted sequences of UTF-16 code
units.

Test Plan:
Create `words` via:

```shell
$ xxd -r >words
00000000: f090 9082 f090 b084 0af0 9090 84f0 90b0  ................
00000010: 820a f090 b084 f090 9082 0af0 90b0 82f0  ................
00000020: 9090 840a 7061 7374 656c 0a73 7461 706c  ....pastel.stapl
00000030: 650a 706c 6174 6573 0a                   e.plates.
```

Run:

```shell
javac src/effectivejava/chapter7/item45/anagrams/*.java
for k in Hybrid Iterative Stream; do \
    java -cp src "effectivejava.chapter7.item45.anagrams.${k}Anagrams" \
            words 1
    printf '%s\n' ---
done
```

Before this patch, the output (for each program) has one group of 3 and
one group of 4. After this patch, the output has one group of 3 and two
groups of 2, which is correct.

wchargin-branch: fix-alphabetize